backend/requirements.txt
fastapi
uvicorn>=0.22.0
websockets<12.0
psycopg2-binary
sqlalchemy
pydantic>=2.5
pydantic-settings>=2.0
passlib
email-validator
confluent-kafka==1.9.2
kafka-python
redis>=4.3.0
fastapi-limiter
aiokafka
finnhub-python
httpx
python-dotenv
asyncpg
boto3==1.26.0
